requirements.txt
beautifulsoup4==4.9.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
html5lib==1.1
idna==2.10
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
pycparser==2.20
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tinycss2==1.0.2
urllib3==1.25.10
WeasyPrint==51
webencodings==0.5.1
